Extract generic Volley calls into VolleyUtil
