--- app/src/main/kotlin/com/github/oryanmat/trellowidget/data/remote/NetworkQueue.kt
@@ -1,17 +1,0 @@
-package com.github.oryanmat.trellowidget.data.remote
-
-import android.content.Context
-import com.android.volley.Request
-import com.android.volley.RequestQueue
-import com.android.volley.toolbox.Volley
-
-class NetworkQueue(appContext: Context) {
-
-    private val requestQueue: RequestQueue by lazy {
-        Volley.newRequestQueue(appContext.applicationContext)
-    }
-
-    fun <T> addToRequestQueue(req: Request<T>) {
-        requestQueue.add(req)
-    }
-}
